+++ tests/test_app/TestApp/Application.php
@@ -1,0 +1,57 @@
+<?php
+/**
+ * This file is part of cakephp-database-backup.
+ *
+ * Licensed under The MIT License
+ * For full copyright and license information, please see the LICENSE.txt
+ * Redistributions of files must retain the above copyright notice.
+ *
+ * @copyright   Copyright (c) Mirko Pagliai
+ * @link        https://github.com/mirko-pagliai/cakephp-database-backup
+ * @license     https://opensource.org/licenses/mit-license.php MIT License
+ * @since       2.6.0
+ */
+namespace App;
+
+use Cake\Http\BaseApplication;
+use Cake\Routing\Middleware\RoutingMiddleware;
+use DatabaseBackup\Plugin as DatabaseBackup;
+
+/**
+ * Application setup class.
+ *
+ * This defines the bootstrapping logic and middleware layers you
+ * want to use in your application.
+ */
+class Application extends BaseApplication
+{
+    /**
+     * Load all the application configuration and bootstrap logic
+     */
+    public function bootstrap()
+    {
+        $this->addPlugin(DatabaseBackup::class);
+    }
+
+    /**
+     * Define the HTTP middleware layers for an application
+     * @param \Cake\Http\MiddlewareQueue $middlewareQueue The middleware queue to set in your App Class
+     * @return \Cake\Http\MiddlewareQueue
+     */
+    public function middleware($middlewareQueue)
+    {
+        $middlewareQueue->add(new RoutingMiddleware($this));
+
+        return $middlewareQueue;
+    }
+
+    /**
+     * Define the routes for an application
+     * @param \Cake\Routing\RouteBuilder $routes A route builder to add routes into
+     * @return void
+     */
+    public function routes($routes)
+    {
+        //Do nothing
+    }
+}